+++ selfdrive/car/gm/values.py
@@ -4,6 +4,33 @@ from cereal import car
 from selfdrive.car import dbc_dict
 Ecu = car.CarParams.Ecu
 
+class CarControllerParams():
+  def __init__(self):
+    self.STEER_MAX = 300
+    self.STEER_STEP = 2              # how often we update the steer cmd
+    self.STEER_DELTA_UP = 7          # ~0.75s time to peak torque (255/50hz/0.75s)
+    self.STEER_DELTA_DOWN = 17       # ~0.3s from peak torque to zero
+    self.MIN_STEER_SPEED = 3.
+    self.STEER_DRIVER_ALLOWANCE = 50   # allowed driver torque before start limiting
+    self.STEER_DRIVER_MULTIPLIER = 4   # weight driver torque heavily
+    self.STEER_DRIVER_FACTOR = 100     # from dbc
+    self.NEAR_STOP_BRAKE_PHASE = 0.5  # m/s, more aggressive braking near full stop
+
+    # Takes case of "Service Adaptive Cruise" and "Service Front Camera"
+    # dashboard messages.
+    self.ADAS_KEEPALIVE_STEP = 100
+    self.CAMERA_KEEPALIVE_STEP = 100
+
+    # pedal lookups, only for Volt
+    MAX_GAS = 3072              # Only a safety limit
+    ZERO_GAS = 2048
+    MAX_BRAKE = 350             # Should be around 3.5m/s^2, including regen
+    self.MAX_ACC_REGEN = 1404  # ACC Regen braking is slightly less powerful than max regen paddle
+    self.GAS_LOOKUP_BP = [-0.25, 0., 0.5]
+    self.GAS_LOOKUP_V = [self.MAX_ACC_REGEN, ZERO_GAS, MAX_GAS]
+    self.BRAKE_LOOKUP_BP = [-1., -0.25]
+    self.BRAKE_LOOKUP_V = [MAX_BRAKE, 0]
+
 class CAR:
   HOLDEN_ASTRA = "HOLDEN ASTRA RS-V BK 2017"
   VOLT = "CHEVROLET VOLT PREMIER 2017"
