+++ selfdrive/car/mazda/carcontroller.py
@@ -1,5 +1,5 @@
 from selfdrive.car.mazda import mazdacan
-from selfdrive.car.mazda.values import SteerLimitParams, Buttons
+from selfdrive.car.mazda.values import CarControllerParams, Buttons
 from opendbc.can.packer import CANPacker
 from selfdrive.car import apply_std_steer_torque_limits
 
@@ -18,9 +18,9 @@ class CarController():
 
     if enabled:
       # calculate steer and also set limits due to driver torque
-      new_steer = int(round(actuators.steer * SteerLimitParams.STEER_MAX))
+      new_steer = int(round(actuators.steer * CarControllerParams.STEER_MAX))
       apply_steer = apply_std_steer_torque_limits(new_steer, self.apply_steer_last,
-                                                  CS.out.steeringTorque, SteerLimitParams)
+                                                  CS.out.steeringTorque, CarControllerParams)
       self.steer_rate_limited = new_steer != apply_steer
 
       if CS.out.standstill and frame % 20 == 0:
